Bugfixes and tournament improvements

+++ TCGUABot/Telegram/CallbackHandlers/TourneyCallbackHandler.cs
@@ -23,36 +23,41 @@ namespace TCGUABot.Models.CallbackHandlers
             var name = dataArray[0];
             var tourneyId = dataArray[1];
 
-            var tUser = query.From;
-
-            Helpers.TelegramUtil.AddUser(tUser, context);
-
-            var tourney = context.Tournaments.FirstOrDefault(t => t.Id == tourneyId);
-
-            if (context.TournamentUserPairs.Any(p => p.TournamentId == tourneyId))
+            if (tourneyId != "refresh")
             {
-                if (!context.TournamentUserPairs.Any(p => p.PlayerTelegramId == query.From.Id && p.TournamentId == tourneyId))
+
+                var tUser = query.From;
+
+                Helpers.TelegramUtil.AddUser(tUser, context);
+
+                var tourney = context.Tournaments.FirstOrDefault(t => t.Id == tourneyId);
+                if (tourney != null)
                 {
-                    var firstPId = context.UserLogins.FirstOrDefault(l => l.ProviderKey == query.From.Id.ToString());
-                    var playerId = firstPId == null ? "0" : firstPId.UserId;
-                    context.TournamentUserPairs.Add(new TournamentUserPair() { PlayerTelegramId = query.From.Id, DeckId = "", PlayerId = playerId, TournamentId = tourneyId });
-                    context.SaveChanges();
-                }
-                else
-                {
-                    var userPair = context.TournamentUserPairs.FirstOrDefault(u => u.PlayerTelegramId == query.From.Id && u.TournamentId == tourneyId);
-                    context.TournamentUserPairs.Remove(userPair);
-                    context.SaveChanges();
+                    if (context.TournamentUserPairs.Any(p => p.TournamentId == tourneyId))
+                    {
+                        if (!context.TournamentUserPairs.Any(p => p.PlayerTelegramId == query.From.Id && p.TournamentId == tourneyId))
+                        {
+                            var firstPId = context.UserLogins.FirstOrDefault(l => l.ProviderKey == query.From.Id.ToString());
+                            var playerId = firstPId == null ? "0" : firstPId.UserId;
+                            context.TournamentUserPairs.Add(new TournamentUserPair() { PlayerTelegramId = query.From.Id, DeckId = "", PlayerId = playerId, TournamentId = tourneyId });
+                            context.SaveChanges();
+                        }
+                        else
+                        {
+                            var userPair = context.TournamentUserPairs.FirstOrDefault(u => u.PlayerTelegramId == query.From.Id && u.TournamentId == tourneyId);
+                            context.TournamentUserPairs.Remove(userPair);
+                            context.SaveChanges();
+                        }
+                    }
+                    else
+                    {
+                        var firstPId = context.UserLogins.FirstOrDefault(l => l.ProviderKey == query.From.Id.ToString());
+                        var playerId = firstPId == null ? "0" : firstPId.UserId;
+                        context.TournamentUserPairs.Add(new TournamentUserPair() { PlayerTelegramId = query.From.Id, DeckId = "", PlayerId = playerId, TournamentId = tourneyId });
+                        context.SaveChanges();
+                    }
                 }
             }
-            else
-            {
-                var firstPId = context.UserLogins.FirstOrDefault(l => l.ProviderKey == query.From.Id.ToString());
-                var playerId = firstPId == null ? "0" : firstPId.UserId;
-                context.TournamentUserPairs.Add(new TournamentUserPair() { PlayerTelegramId = query.From.Id, DeckId = "", PlayerId = playerId, TournamentId = tourneyId });
-                context.SaveChanges();
-            }
-
 
             var TList = context.Tournaments.Where(t => t.IsClosed == false).ToList();
             var chatId = query.Message.Chat.Id;
